refactor: 타입 분리 및 배럴 export 설정(#81)
