+++ src/api/member/index.ts
@@ -1,15 +1,5 @@
-export type { NicknameCheckData } from "./nicknameCheck.api";
 export { nicknameCheck } from "./nicknameCheck.api";
 
-export type {
-  PhoneVerifyRequestData,
-  PhoneVerifyRequestReq,
-  PhoneVerifyPurpose,
-} from "./requestPhoneVerification.api";
 export { requestPhoneVerification } from "./requestPhoneVerification.api";
 
-export type {
-  PhoneVerifyConfirmData,
-  PhoneVerifyConfirmReq,
-} from "./confirmPhoneVerification.api";
 export { confirmPhoneVerification } from "./confirmPhoneVerification.api";
