+++ src/api/member/requestPhoneVerification.api.ts
@@ -1,17 +1,9 @@
 import { axiosInstance } from "@/lib/axiosInstance";
 import type { ApiResponse } from "@/types/common/apiResponse";
-
-export type PhoneVerifyPurpose = "SIGNUP" | "MY_PAGE";
-
-export interface PhoneVerifyRequestReq {
-  phone: string;
-  purpose: PhoneVerifyPurpose;
-}
-
-export interface PhoneVerifyRequestData {
-  requestId: string;
-  expiresIn: number;
-}
+import type {
+  PhoneVerifyRequestData,
+  PhoneVerifyRequestReq,
+} from "@/types/member";
 
 // 휴대폰 인증번호 요청
 export async function requestPhoneVerification(
