+++ src/hooks/auth/login/useOauth.ts
@@ -1,8 +1,8 @@
-import {
-  oauth,
-  type SocialoginCompleteData,
-  type SocialoginCompleteReq,
-} from "@/api/auth";
+import { oauth } from "@/api/auth";
+import type {
+  SocialoginCompleteData,
+  SocialoginCompleteReq,
+} from "@/types/auth";
 import type { ApiResponse } from "@/types/common/apiResponse";
 
 import { useMutation, type UseMutationOptions } from "@tanstack/react-query";
