+++ src/types/auth/index.ts
@@ -1,0 +1,11 @@
+export type {
+  SocialProfile,
+  SocialoginCompleteData,
+  SocialoginCompleteReq,
+} from "./oAuth";
+
+export type {
+  AgreementItem,
+  SocialSignupCompleteData,
+  SocialSignupCompleteReq,
+} from "./signUp";
