+++ src/types/member/index.ts
@@ -1,0 +1,9 @@
+export type { NicknameCheckData } from "./nickname";
+
+export type {
+  PhoneVerifyConfirmData,
+  PhoneVerifyConfirmReq,
+  PhoneVerifyPurpose,
+  PhoneVerifyRequestData,
+  PhoneVerifyRequestReq,
+} from "./phone";
