+++ src/types/member/nickname.ts
@@ -1,0 +1,4 @@
+export interface NicknameCheckData {
+  nickname: string;
+  available: boolean;
+}
